+++ src/lib.rs
@@ -37,7 +37,7 @@ mod response;
     feature = "token"
 ))]
 pub mod auth;
-#[cfg(any(feature = "kv1", feature = "kv2"))]
+#[cfg(any(feature = "kv1", feature = "kv2", feature = "pki", feature = "transit"))]
 pub mod secrets;
 #[cfg(feature = "sys")]
 pub mod sys;