+++ src/secrets/mod.rs
@@ -4,5 +4,7 @@
 pub mod kv1;
 #[cfg(feature = "kv2")]
 pub mod kv2;
+#[cfg(feature = "pki")]
+pub mod pki;
 #[cfg(feature = "transit")]
 pub mod transit;